+++ R/fscores.internal.R
@@ -150,7 +150,7 @@ setMethod(
 	        colSums(probs)
 	    }
 
-        if(plausible.draws > 0){
+        if(plausible.draws > 0 && is.null(response.pattern)){
             if(plausible.type == 'MH'){
                 dots <- list(...)
                 technical <- if(!is.null(dots$technical)) dots$technical else list()
@@ -220,8 +220,10 @@ setMethod(
                                method=method, quadpts=quadpts, verbose=FALSE, full.scores.SE=TRUE,
                                response.pattern=NULL, return.acov=return.acov, theta_lim=theta_lim,
                                MI=MI, mean=gmean, cov=gcov, custom_den=custom_den, QMC=QMC,
-                               custom_theta=custom_theta,
-                               start=start, use_dentype_estimate=use_dentype_estimate, ...)
+                               custom_theta=custom_theta, plausible.draws=plausible.draws,
+                               plausible.type=plausible.type, start=start,
+                               use_dentype_estimate=use_dentype_estimate, ...)
+                if(plausible.draws > 0) return(ret)
                 if(return.acov) return(ret)
                 if(append_response.pattern) ret <- cbind(response.pattern, ret)
             } else {